Rename card presenters and views
