+++ src/views/index.js
@@ -1,11 +1,11 @@
 import BankView from "./bankView.js"
 import BoardView from "./boardView.js"
-import CardView from "./cardView"
+import TranslationCardView from "./translationCardView"
 import AddBoardView from "./addBoardView"
 import SidebarView from "./sidebarView"
-import CardInfoView from "./cardInfoView"
+import CardModalView from "./cardModalView"
 
 export {
-        BankView, BoardView, CardView, AddBoardView,
-        SidebarView, CardInfoView
+        BankView, BoardView, TranslationCardView, AddBoardView,
+        SidebarView, CardModalView
 }
